updated Readme (#49)
* updated Readme

created one readme without examples that would need to be updated and a clear statement about the current stable version. 

This Readme could be used for all versions without and it will be consistent on Docker UP, no matter what was the last push. As Docker Hub only takes the latest push readme and has no option to mark one as stable.

* Wording / Fix