fix wrong check order
the if-elif was done in the wrong order. Now `http_bind_adresse` is the first check and the `check_url` might be created if the variable is given. The elif can now overwrite `check_url` what is done if `http_publish_uri`is given. What makes the script behave like described.