updated health_check (#68)
* updated health_check

reworked the health_check that should now work in all conditions. 

The only uncovered case is when someone gives HTTP_PUBLISH_URI a wildcard address what will make Graylog take a "the first" interface. IMHO that should not happen in Docker environments.

* cosmetical changes

* fix wrong check order

the if-elif was done in the wrong order. Now `http_bind_adresse` is the first check and the `check_url` might be created if the variable is given. The elif can now overwrite `check_url` what is done if `http_publish_uri`is given. What makes the script behave like described.