multiple fixes for health check and added feature (#107)
This will fix the issue that the `http_publish_uri`can't be used for the health_check. So the fallback is localhost.

In every normal setup that should not be the case but having an edge case setup might have this impact.

FIX: https://github.com/Graylog2/graylog-docker/issues/101

This removes the protocol from the `http_publish_uri` from configuration files. 

Not the smartes way to do this, but currently this is working and we might want to rework that in the future.

FIX: https://github.com/Graylog2/graylog-docker/issues/102


Taking the proposed changes from https://github.com/Graylog2/graylog-docker/issues/103

and added this to the entry point. Will also add this to the documentation that this is know behaviour.

CLOSE https://github.com/Graylog2/graylog-docker/issues/103