Refactor forwarder image
Use the release tarball instead of the DEB package to make it possible
to build pre-release images. We don't publish forwarder operating system
packages for pre-releases.
This also decouples the image from the operating system package
revision.

All the paths stay the same so it should be backward compatible to the
image that was using the DEB package for installation.